add 30s default value for timeout

+++ controller/operation/stop.js
@@ -11,15 +11,23 @@ command {
 module.exports = function stop(config, command) {
   return new Promise((resolve, reject) => {
 
-    const timeout = +command.timeout || null
+    let timeout = 1000 * 30
 
-    const killed = Worker.workerList.filter(worker => {
-      return worker.name === command.app
-    }).map(worker => {
-      if(worker.kill(timeout)) {
-        return worker.once('exit')
+    if(command.timeout) {
+      if(isNaN(+command.timeout)) {
+        timeout = null
+      } else {
+        timeout = +command.timeout
       }
-    })
+    }
+
+    const killed = Worker.workerList.map(worker => {
+      if(worker.name === command.app) {
+        if(worker.kill(timeout)) {
+          return worker.once('exit')
+        }
+      }
+    }).filter(p => p)
 
     Promise.all(killed).then(() => {
       let i = config.apps.findIndex(app => app.name === command.app)
